refactor: move DataProvenanceMixin to data_provenance.py and refactor for class-level usage
- Relocated DataProvenanceMixin from base.py to data_provenance.py for better modularity and provenance management.
- Refactored mixin to use cls in @declared_attr for proper class-level relationship definition.
